+++ app/src/main/kotlin/com/d4rk/android/apps/apptoolkit/app/main/ui/components/navigation/NavigationDrawer.kt
@@ -6,7 +6,6 @@ import androidx.compose.material3.DrawerValue
 import androidx.compose.material3.ModalDrawerSheet
 import androidx.compose.material3.ModalNavigationDrawer
 import androidx.compose.material3.rememberDrawerState
-import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.rememberCoroutineScope
@@ -15,12 +14,17 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import com.d4rk.android.apps.apptoolkit.app.main.ui.MainScaffoldContent
 import com.d4rk.android.apps.apptoolkit.app.main.ui.states.MainUiState
+import com.d4rk.android.apps.apptoolkit.app.main.utils.constants.AppNavKey
 import com.d4rk.android.libs.apptoolkit.app.main.domain.model.BottomBarItem
 import com.d4rk.android.libs.apptoolkit.app.main.ui.components.dialogs.ChangelogDialog
 import com.d4rk.android.libs.apptoolkit.app.main.ui.components.navigation.NavigationDrawerItemContent
+import com.d4rk.android.libs.apptoolkit.app.main.ui.navigation.handleNavigationItemClick
 import com.d4rk.android.libs.apptoolkit.core.domain.model.navigation.NavigationDrawerItem
 import com.d4rk.android.libs.apptoolkit.core.ui.components.modifiers.hapticDrawerSwipe
 import com.d4rk.android.libs.apptoolkit.core.ui.components.spacers.LargeVerticalSpacer
+import com.d4rk.android.libs.apptoolkit.core.ui.navigation.NavigationState
+import com.d4rk.android.libs.apptoolkit.core.ui.navigation.Navigator
+import com.d4rk.android.libs.apptoolkit.core.utils.window.AppWindowWidthSizeClass
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.coroutines.CoroutineScope
 import org.koin.compose.koinInject
@@ -29,8 +33,10 @@ import org.koin.core.qualifier.named
 @Composable
 fun NavigationDrawer(
     uiState: MainUiState,
-    windowWidthSizeClass: WindowWidthSizeClass,
-    bottomItems: ImmutableList<BottomBarItem>,
+    windowWidthSizeClass: AppWindowWidthSizeClass,
+    bottomItems: ImmutableList<BottomBarItem<AppNavKey>>,
+    navigationState: NavigationState<AppNavKey>,
+    navigator: Navigator<AppNavKey>,
 ) {
     val drawerState: DrawerState = rememberDrawerState(initialValue = DrawerValue.Closed)
     val coroutineScope: CoroutineScope = rememberCoroutineScope()
@@ -66,6 +72,8 @@ fun NavigationDrawer(
             drawerState = drawerState,
             windowWidthSizeClass = windowWidthSizeClass,
             bottomItems = bottomItems,
+            navigationState = navigationState,
+            navigator = navigator,
         )
     }