+++ app/src/test/kotlin/com/d4rk/android/apps/apptoolkit/app/main/ui/components/navigation/AppNavigationHostTest.kt
@@ -1,7 +1,11 @@
 package com.d4rk.android.apps.apptoolkit.app.main.ui.components.navigation
 
+import com.d4rk.android.apps.apptoolkit.app.main.utils.constants.AppsListRoute
+import com.d4rk.android.apps.apptoolkit.app.main.utils.constants.FavoriteAppsRoute
 import com.d4rk.android.apps.apptoolkit.app.main.utils.constants.NavigationRoutes
+import com.d4rk.android.apps.apptoolkit.app.main.utils.constants.toNavKeyOrDefault
 import com.d4rk.android.apps.apptoolkit.core.data.datastore.DataStore
+import com.d4rk.android.libs.apptoolkit.data.datastore.startupDestinationFlow
 import io.mockk.clearAllMocks
 import io.mockk.every
 import io.mockk.mockk
@@ -30,10 +34,13 @@ class AppNavigationHostTest {
             ""
         )
 
-        val startDestination = dataStore.startupDestinationFlow().first()
+        val startDestination = dataStore.startupDestinationFlow(
+            defaultRoute = NavigationRoutes.ROUTE_APPS_LIST,
+            mapToKey = { route -> route.toNavKeyOrDefault() }
+        ).first()
 
-        assertEquals(NavigationRoutes.ROUTE_APPS_LIST, startDestination)
-        verify(exactly = 1) { dataStore.getStartupPage(default = NavigationRoutes.ROUTE_APPS_LIST) }
+        assertEquals(AppsListRoute, startDestination)
+        verify(exactly = 1) { dataStore.getStartupPage(default = NavigationRoutes.ROUTE_APPS_LIST) } // FIXME: Flow is constructed but not used
     }
 
     @Test
@@ -42,9 +49,12 @@ class AppNavigationHostTest {
             dataStore.getStartupPage(default = NavigationRoutes.ROUTE_APPS_LIST)
         } returns flowOf(NavigationRoutes.ROUTE_FAVORITE_APPS)
 
-        val startDestination = dataStore.startupDestinationFlow().first()
+        val startDestination = dataStore.startupDestinationFlow(
+            defaultRoute = NavigationRoutes.ROUTE_APPS_LIST,
+            mapToKey = { route -> route.toNavKeyOrDefault() }
+        ).first()
 
-        assertEquals(NavigationRoutes.ROUTE_FAVORITE_APPS, startDestination)
-        verify(exactly = 1) { dataStore.getStartupPage(default = NavigationRoutes.ROUTE_APPS_LIST) }
+        assertEquals(FavoriteAppsRoute, startDestination)
+        verify(exactly = 1) { dataStore.getStartupPage(default = NavigationRoutes.ROUTE_APPS_LIST) } // FIXME: Flow is constructed but not used
     }
 }